fx bug

+++ authorization/init.py
@@ -1,10 +1,14 @@
 # -*- coding:utf-8 -*-
+import os
 from datetime import datetime
 from authorization.models import User, Company
 from authorization import db
 
 
 def setup():
+    if 'mysql' in os.environ.get('DB_URL'):
+        return
+
     for c in Company.query.all():
         db.session.delete(c)
     db.session.commit()
